Use smaller offset integers
None of them need to be 64 bits

+++ GameOffsets.h
@@ -5,8 +5,8 @@
 
 #include "Offsets.h"
 
-constexpr inline AnyMemberOffset operator"" _i(unsigned long long n) { return AnyMemberOffset(n); }
-constexpr inline AnyVtableOffset operator"" _v(unsigned long long n) { return AnyVtableOffset(n); }
+constexpr inline AnyMemberOffset operator"" _i(unsigned long long n) { return AnyMemberOffset((uint32_t)n); }
+constexpr inline AnyVtableOffset operator"" _v(unsigned long long n) { return AnyVtableOffset((uint32_t)n); }
 
 #pragma mark Onikakushi, Watanagashi (5.2.2f1)
